Ensure that a paginator exists before calculating page count.
If the application configuration includes a default paginator of `none`
and a `top_level_meta_include_page_count=true`, `show_related_resources`
throws an exception in the Processor as it attempts to calculate the
page count for a `null` processor.

This happens when I want to default to no pagination but on certain
resources want include `page` pagination.